[FIX] refactoring code construction using service package
