+++ bin/main/application.properties
@@ -1,7 +1,7 @@
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mariadb://localhost:3306/bbs?characterEncoding=UTF-8
+spring.datasource.url=jdbc:mysql://localhost:3307/bbs?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=zerobase
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+spring.datasource.password=h2pm2703!123
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql: true
 spring.jwt.secret=zeroBaseJWTPractice
