--- src/main/java/com/personal/bulletinboardsystem/user/entity/UserService.java
@@ -1,5 +1,0 @@
-package com.personal.bulletinboardsystem.user.entity;
-
-public class UserService {
-    
-}
