+++ src/main/java/com/personal/bulletinboardsystem/util/JWTUtils.java
@@ -7,12 +7,10 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class JWTUtils {
-    private static final String KEY = "helloWorld";
-
-    public static String getIssuer(String token) {
-        return JWT.require(Algorithm.HMAC512(KEY))
+    public String getSubject(String token, String jwtSecret) {
+        return JWT.require(Algorithm.HMAC512(jwtSecret))
         .build()
         .verify(token)
-        .getIssuer();
+        .getSubject();
     }
 }
